requirements.txt
asgiref==3.7.2
certifi==2023.11.17
distlib==0.3.7
Django==4.2.7
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
filelock==3.13.1
fpdf==1.7.2
mysqlclient==2.2.0
pipenv==2023.11.15
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6
